pyproject.toml
[tool.poetry]
name = "python-web-07"
version = "0.1.0"
description = ""
authors = ["BogdanKozlov35 <bkozlov35@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
faker = "^25.8.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
